video working

+++ src/Components/Routes/SavedDecks/SavedDecks.js
@@ -1,21 +1,24 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux';
 import axios from 'axios';
-import { updateDecks, updateUserDeckCards } from '../../../ducks/reducer';
+import { updateDecks, updateUserDeckCards, updateDeckName } from '../../../ducks/reducer';
 import UserDeckCards from './UserDeckCards/UserDeckCards';
+import SavedDecksCard from './SavedDecksCard';
 
 
 class SavedDecks extends Component {
     constructor() {
         super();
         this.state = {
-
+            toggle: false,
+            editThis: null
         }
     }
-
+    
     componentDidMount() {
+        
         this.getUserDecks();
-        this.getUserDeckCards();
+        this.getUserCards();
     }
 
     getUserDecks = () => {
@@ -25,7 +28,7 @@ class SavedDecks extends Component {
         })
     }
 
-    getUserDeckCards = (deck_id) => {
+    getUserCards = (deck_id) => {
         console.log('decks id',this.props.decks.id);
         console.log('deck_id =======>', deck_id);
         axios.get(`/api/deckCards/${deck_id && deck_id}`).then(response => {
@@ -35,23 +38,56 @@ class SavedDecks extends Component {
     }
 
     deleteDeck = deck_id => {
+        if(window.confirm('Are you sure you want to delete this deck?')){
         axios.delete(`/api/decks/${deck_id}`).then(response => {
             console.log('delete deck response', response.data)
             this.props.updateDecks(response.data);
             this.getUserDecks();
+        })}
+    }
+
+    updateNameOfDeck = (deck_id, Name) => {
+        console.log('deck_id in update', deck_id);
+        this.toggler();
+        const updatedDeck = {
+
+        }
+        axios.put(`/api/decks/${deck_id}`, {Name: this.props.deckName}).then(response => {
+            this.props.updateDeckName(response.data);
+            this.getUserDecks();
+
+        })
+    }
+
+    toggler = () => {
+        console.log('toggled',new Date, this.state.toggle);
+        this.setState((prevState) => {
+            return {
+                toggle: !prevState.toggle
+            };
         })
     }
 
     render(){
         console.log('userDecks images', this.props.decks)
         const userDecks = this.props.decks.map(deck => {
+            console.log('----deck', deck)
+            // console.groupEnd();
         return (
-            <div className={this.props.user ? 'user-deck' : 'hide'}>
-                <button onClick={()=> {this.deleteDeck(deck.id)}} >Del</button>
-                <div className='deck-images'>
-                    <img src={deck.deckimage} onClick={()=> {this.getUserDeckCards(deck.id)}}  />
-                </div>
-                <p>{deck.name}</p>
+            <div>
+               <SavedDecksCard
+                    deck={deck}
+                    deckimage={deck.deckimage}
+                    id={deck.id}
+                    name={deck.name}
+                    user_id={deck.user_id}
+                    updateNameOfDeck={this.updateNameOfDeck}
+                    getUserCards={this.getUserCards}
+                    toggle={this.state.toggle}
+                    toggler={this.toggler}
+                    editThis={this.state.editThis}
+                    handleEdit={this.handleEdit}
+                    />
             </div>
         )
         })
@@ -77,13 +113,15 @@ const mapStateToProps = state => {
     return {
         decks: state.decks,
         userDeckCards: state.userDeckCards,
-        user: state.user
+        user: state.user,
+        deckName: state.deckName
     }
 }
 
 const mapDispatchToProps = {
     updateDecks: updateDecks,
-    updateUserDeckCards: updateUserDeckCards
+    updateUserDeckCards: updateUserDeckCards,
+    updateDeckName: updateDeckName
 }
 
 export default connect(mapStateToProps, mapDispatchToProps)(SavedDecks)
